Add Gemfile, use redcarpet, fix *.html.md issue
@vdavez reported that https://pages.18f.gov/slides/google-apps/ wasn't working
as expected. It turns out that the `markdown_ext: ignoreme` setting in
`_config.yml` was causing the rendered `_slides` pages to show up with the
extension `.html.md`.

Removed this configuration item, switched the Markdown rendering engine from
kramdown to redcarpet, and added a Gemfile for good measure. Confirmed locally
that this resolves the issue at http://localhost:4000/slides/google-apps/#/.