Ensure all slides are .html files, not .md
@vzvenyach noted after #22 that the slides functionality itself wasn't
working. Realized that was because the .md file extensions were causing Jekyll
to render _all_ of the Markdown in the file, whereas reveal.js expects to see
raw Markdown in the HTML <body>. Renaming the files fixes this and restores
slides functionality.

Also took the liberty of updating the README.